crypto-azure-functions/requirements.txt
# requirements.txt

# Core Azure Functions libraries - using specific, compatible versions
azure-functions==1.18.0
azure-functions-durable==1.3.1

# Your other required libraries
requests
azure-storage-blob
pandas
pyarrow
